MySQL / MariaDB - add support concurrent database connections

+++ Json/Fliox.Hub.MySQL/MySQLContainer.cs
@@ -20,21 +20,29 @@ namespace Friflo.Json.Fliox.Hub.MySQL
     {
         private             bool            tableExists;
         public   override   bool            Pretty      { get; }
-        private  readonly   MySQLDatabase   database;
+        private  readonly   MySQLProvider   provider;
         
         internal MySQLContainer(string name, MySQLDatabase database, bool pretty)
             : base(name, database)
         {
-            Pretty          = pretty;
-            this.database   = database;
+            Pretty      = pretty;
+            provider    = database.Provider;
+        }
+        
+        private static MySqlCommand Command (string sql, SyncConnection connection) {
+            return new MySqlCommand(sql, connection.instance as MySqlConnection);
         }
 
-        private async Task<TaskExecuteError> EnsureContainerExists() {
+        private async Task<TaskExecuteError> EnsureContainerExists(SyncContext syncContext) {
             if (tableExists) {
                 return null;
             }
             var sql = $"CREATE TABLE if not exists {name} (id VARCHAR(128) PRIMARY KEY, data JSON);";
-            var result = await MySQLUtils.Execute(database.connection, sql).ConfigureAwait(false);
+            var connection = await syncContext.GetConnection();
+            if (connection.Failed) {
+                return connection.error;
+            }
+            var result = await MySQLUtils.Execute(connection.instance as MySqlConnection, sql).ConfigureAwait(false);
             if (!result.Success) {
                 return result.error;
             }
@@ -43,7 +51,7 @@ namespace Friflo.Json.Fliox.Hub.MySQL
         }
         
         public override async Task<CreateEntitiesResult> CreateEntitiesAsync(CreateEntities command, SyncContext syncContext) {
-            var error = await EnsureContainerExists().ConfigureAwait(false);
+            var error = await EnsureContainerExists(syncContext).ConfigureAwait(false);
             if (error != null) {
                 return new CreateEntitiesResult { Error = error };
             }
@@ -53,8 +61,12 @@ namespace Friflo.Json.Fliox.Hub.MySQL
             var sql = new StringBuilder();
             sql.Append($"INSERT INTO {name} (id,data) VALUES\n");
             SQLUtils.AppendValuesSQL(sql, command.entities);
-            using var cmd = new MySqlCommand(sql.ToString(), database.connection);
+            var connection = await syncContext.GetConnection();
+            if (connection.Failed) {
+                return new CreateEntitiesResult { Error = connection.error };
+            }
             try {
+                using var cmd = Command(sql.ToString(), connection);
                 await cmd.ExecuteNonQueryAsync().ConfigureAwait(false);
             } catch (Exception e) {
                 return new CreateEntitiesResult { Error = DatabaseError(e.Message) };    
@@ -63,7 +75,7 @@ namespace Friflo.Json.Fliox.Hub.MySQL
         }
         
         public override async Task<UpsertEntitiesResult> UpsertEntitiesAsync(UpsertEntities command, SyncContext syncContext) {
-            var error = await EnsureContainerExists().ConfigureAwait(false);
+            var error = await EnsureContainerExists(syncContext).ConfigureAwait(false);
             if (error != null) {
                 return new UpsertEntitiesResult { Error = error };
             }
@@ -73,14 +85,18 @@ namespace Friflo.Json.Fliox.Hub.MySQL
             var sql = new StringBuilder();
             sql.Append($"REPLACE INTO {name} (id,data) VALUES\n");
             SQLUtils.AppendValuesSQL(sql, command.entities);
-            using var cmd = new MySqlCommand(sql.ToString(), database.connection);
+            var connection = await syncContext.GetConnection();
+            if (connection.Failed) {
+                return new UpsertEntitiesResult { Error = connection.error };
+            }
+            using var cmd = Command(sql.ToString(), connection);
             await cmd.ExecuteNonQueryAsync().ConfigureAwait(false);
 
             return new UpsertEntitiesResult();
         }
 
         public override async Task<ReadEntitiesResult> ReadEntitiesAsync(ReadEntities command, SyncContext syncContext) {
-            var error = await EnsureContainerExists().ConfigureAwait(false);
+            var error = await EnsureContainerExists(syncContext).ConfigureAwait(false);
             if (error != null) {
                 return new ReadEntitiesResult { Error = error };
             }
@@ -88,20 +104,28 @@ namespace Friflo.Json.Fliox.Hub.MySQL
             var sql = new StringBuilder();
             sql.Append($"SELECT id, data FROM {name} WHERE id in\n");
             SQLUtils.AppendKeysSQL(sql, ids);
-            using var cmd   = new MySqlCommand(sql.ToString(), database.connection);
+            var connection = await syncContext.GetConnection();
+            if (connection.Failed) {
+                return new ReadEntitiesResult { Error = connection.error };
+            }
+            using var cmd = Command(sql.ToString(), connection);
             return await SQLUtils.ReadEntities(cmd, command).ConfigureAwait(false);
         }
 
         public override async Task<QueryEntitiesResult> QueryEntitiesAsync(QueryEntities command, SyncContext syncContext) {
-            var error = await EnsureContainerExists().ConfigureAwait(false);
+            var error = await EnsureContainerExists(syncContext).ConfigureAwait(false);
             if (error != null) {
                 return new QueryEntitiesResult { Error = error };
             }
-            var filter  = command.GetFilter();
-            var where   = filter.IsTrue ? "TRUE" : filter.MySQLFilter(database.Provider);
-            var sql     = SQLUtils.QueryEntitiesSQL(command, name, where);
+            var filter      = command.GetFilter();
+            var where       = filter.IsTrue ? "TRUE" : filter.MySQLFilter(provider);
+            var sql         = SQLUtils.QueryEntitiesSQL(command, name, where);
+            var connection  = await syncContext.GetConnection();
+            if (connection.Failed) {
+                return new QueryEntitiesResult { Error = connection.error };
+            }
             try {
-                using var cmd = new MySqlCommand(sql, database.connection);
+                using var cmd = Command(sql, connection);
                 return await SQLUtils.QueryEntities(cmd, command, sql).ConfigureAwait(false);
             }
             catch (MySqlException e) {
@@ -112,24 +136,31 @@ namespace Friflo.Json.Fliox.Hub.MySQL
         public override async Task<AggregateEntitiesResult> AggregateEntitiesAsync (AggregateEntities command, SyncContext syncContext) {
             if (command.type == AggregateType.count) {
                 var filter  = command.GetFilter();
-                var where   = filter.IsTrue ? "" : $" WHERE {filter.MySQLFilter(database.Provider)}";
+                var where   = filter.IsTrue ? "" : $" WHERE {filter.MySQLFilter(provider)}";
                 var sql     = $"SELECT COUNT(*) from {name}{where}";
-                var result  = await MySQLUtils.Execute(database.connection, sql).ConfigureAwait(false);
+                var connection = await syncContext.GetConnection();
+                if (connection.Failed) {
+                    return new AggregateEntitiesResult { Error = connection.error };
+                }
+                var result  = await MySQLUtils.Execute(connection.instance as MySqlConnection, sql).ConfigureAwait(false);
                 if (!result.Success) { return new AggregateEntitiesResult { Error = result.error }; }
                 return new AggregateEntitiesResult { value = (long)result.value };
             }
             throw new NotImplementedException();
         }
-        
-       
+
         public override async Task<DeleteEntitiesResult> DeleteEntitiesAsync(DeleteEntities command, SyncContext syncContext) {
-            var error = await EnsureContainerExists().ConfigureAwait(false);
+            var error = await EnsureContainerExists(syncContext).ConfigureAwait(false);
             if (error != null) {
                 return new DeleteEntitiesResult { Error = error };
             }
+            var connection = await syncContext.GetConnection();
+            if (connection.Failed) {
+                return new DeleteEntitiesResult { Error = connection.error };
+            }
             if (command.all == true) {
                 var sql = $"DELETE from {name}";
-                var result = await MySQLUtils.Execute(database.connection, sql).ConfigureAwait(false);
+                var result = await MySQLUtils.Execute(connection.instance as MySqlConnection, sql).ConfigureAwait(false);
                 if (!result.Success) { return new DeleteEntitiesResult { Error = result.error }; }
                 return new DeleteEntitiesResult();    
             } else {
@@ -137,7 +168,7 @@ namespace Friflo.Json.Fliox.Hub.MySQL
                 sql.Append($"DELETE FROM  {name} WHERE id in\n");
                 
                 SQLUtils.AppendKeysSQL(sql, command.ids);
-                using var cmd = new MySqlCommand(sql.ToString(), database.connection);
+                using var cmd = Command(sql.ToString(), connection);
                 await cmd.ExecuteNonQueryAsync().ConfigureAwait(false);
                 return new DeleteEntitiesResult();
             }
